% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{ACTIVE_SET_THRESHOLD}
\alias{ACTIVE_SET_THRESHOLD}
\title{Active set threshold (matches C++ ACTIVE_SET_THRESHOLD)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
ACTIVE_SET_THRESHOLD
}
\description{
Active set threshold (matches C++ ACTIVE_SET_THRESHOLD)
}
\keyword{datasets}
