% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{calculate_site_probabilities}
\alias{calculate_site_probabilities}
\title{Calculate site membership probabilities
Following eq:site_membership in main.tex: P(R=t|X) and P(R=s_j,A=a|X) using multinomial logistic model}
\usage{
calculate_site_probabilities(Z, gamma_params, K)
}
\arguments{
\item{Z}{covariate matrix (X or X_dagger)}

\item{gamma_params}{list of γ (site/treatment model) parameters}

\item{K}{number of source sites}
}
\value{
list with probabilities for each site and treatment combination
}
\description{
Calculate site membership probabilities
Following eq:site_membership in main.tex: P(R=t|X) and P(R=s_j,A=a|X) using multinomial logistic model
}
