% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkpoint.R
\name{check_preempt_signal}
\alias{check_preempt_signal}
\title{Check if a SLURM preemption signal was received}
\usage{
check_preempt_signal(preempt_signal_file)
}
\arguments{
\item{preempt_signal_file}{Path to the preemption signal file.}
}
\value{
TRUE if the signal file exists.
}
\description{
Check if a SLURM preemption signal was received
}
