% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{CHECKPOINT_SETTING_INTERVAL}
\alias{CHECKPOINT_SETTING_INTERVAL}
\title{Checkpoint save interval: save every N parameter settings}
\format{
An object of class \code{integer} of length 1.
}
\usage{
CHECKPOINT_SETTING_INTERVAL
}
\description{
Checkpoint save interval: save every N parameter settings
}
\keyword{datasets}
