% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{CHECKPOINT_SIM_INTERVAL}
\alias{CHECKPOINT_SIM_INTERVAL}
\title{Checkpoint save interval: save every N simulations within a setting}
\format{
An object of class \code{integer} of length 1.
}
\usage{
CHECKPOINT_SIM_INTERVAL
}
\description{
Checkpoint save interval: save every N simulations within a setting
}
\keyword{datasets}
