% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkpoint.R
\name{cleanup_checkpoint}
\alias{cleanup_checkpoint}
\title{Clean up all checkpoint artifacts after successful completion}
\usage{
cleanup_checkpoint(checkpoint_file, preempt_signal_file, saved_signal_file)
}
\arguments{
\item{checkpoint_file}{Path to checkpoint RData file.}

\item{preempt_signal_file}{Path to preemption signal file.}

\item{saved_signal_file}{Path to saved-signal file.}
}
\description{
Clean up all checkpoint artifacts after successful completion
}
