% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{DIVISION_FLOOR}
\alias{DIVISION_FLOOR}
\title{Division floor to prevent divide-by-zero (denominator clipping)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
DIVISION_FLOOR
}
\description{
Division floor to prevent divide-by-zero (denominator clipping)
}
\keyword{datasets}
