% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{DR_WEIGHT_LOWER}
\alias{DR_WEIGHT_LOWER}
\title{Density ratio weight clipping bounds}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
DR_WEIGHT_LOWER
}
\description{
Density ratio weight clipping bounds
}
\keyword{datasets}
