% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{EPSILON_DEFAULT}
\alias{EPSILON_DEFAULT}
\title{Default epsilon for numerical stability operations}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
EPSILON_DEFAULT
}
\description{
Default epsilon for numerical stability operations
}
\keyword{datasets}
