% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{ESTIMATE_MAX}
\alias{ESTIMATE_MAX}
\title{Maximum absolute value for estimates (clipping bound)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
ESTIMATE_MAX
}
\description{
Maximum absolute value for estimates (clipping bound)
}
\keyword{datasets}
