% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{FACE_ATE_TARGET}
\alias{FACE_ATE_TARGET}
\title{True ATE for the target population in FACE paper DGP, continuous outcome
(Δ_T = 3.0, the conditional mean shift between arms).}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
FACE_ATE_TARGET
}
\description{
True ATE for the target population in FACE paper DGP, continuous outcome
(Δ_T = 3.0, the conditional mean shift between arms).
}
\keyword{datasets}
