% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{FACE_KAPPA}
\alias{FACE_KAPPA}
\title{Shared location parameter κ for the skewed-normal covariate distribution.
All sites use the same κ (midpoint of the (0.10, 0.15) range in the paper).}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
FACE_KAPPA
}
\description{
Shared location parameter κ for the skewed-normal covariate distribution.
All sites use the same κ (midpoint of the (0.10, 0.15) range in the paper).
}
\keyword{datasets}
