% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{FACE_NOISE_SD}
\alias{FACE_NOISE_SD}
\title{Noise standard deviation for FACE paper outcomes: 2√5 ≈ 4.47}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
FACE_NOISE_SD
}
\description{
Noise standard deviation for FACE paper outcomes: 2√5 ≈ 4.47
}
\keyword{datasets}
