% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{FACE_NU_SOURCE_MAX}
\alias{FACE_NU_SOURCE_MAX}
\title{Maximum skewness parameter ν for source-site covariates (paper: ν ∈ [0, 0.2]).
Source sites receive equally-spaced ν values between ν_max/K and ν_max.}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
FACE_NU_SOURCE_MAX
}
\description{
Maximum skewness parameter ν for source-site covariates (paper: ν ∈ [0, 0.2]).
Source sites receive equally-spaced ν values between ν_max/K and ν_max.
}
\keyword{datasets}
