% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{FACE_SIGNAL_COORDINATES}
\alias{FACE_SIGNAL_COORDINATES}
\title{Number of leading covariates that carry the signal (non-zero outcome and
propensity coefficients) and, under misspecification, the Kang--Schafer-style
transforms of \code{.face_transformed_coordinates()}.}
\format{
An object of class \code{integer} of length 1.
}
\usage{
FACE_SIGNAL_COORDINATES
}
\description{
Number of leading covariates that carry the signal (non-zero outcome and
propensity coefficients) and, under misspecification, the Kang--Schafer-style
transforms of \code{.face_transformed_coordinates()}.
}
\keyword{datasets}
