% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators_helpers.R
\name{fit_glm_mle}
\alias{fit_glm_mle}
\title{Fit unpenalized GLM MLE with intercept}
\usage{
fit_glm_mle(X, y, family = "binomial")
}
\arguments{
\item{X}{covariate matrix (n x p)}

\item{y}{response vector}

\item{family}{GLM family ("binomial" or "gaussian")}
}
\value{
list with coefficients and fitted values on response scale
}
\description{
Fit unpenalized GLM MLE with intercept
}
