% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators_helpers.R
\name{fit_logit_mle}
\alias{fit_logit_mle}
\title{Fit unpenalized logit MLE with intercept}
\usage{
fit_logit_mle(X, y)
}
\arguments{
\item{X}{covariate matrix (n x p)}

\item{y}{binary response}
}
\value{
list with coefficients and fitted probabilities
}
\description{
Fit unpenalized logit MLE with intercept
}
