% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{generate_base_covariates}
\alias{generate_base_covariates}
\title{Generate base covariates}
\usage{
generate_base_covariates(n, p = 4)
}
\arguments{
\item{n}{number of observations}

\item{p}{number of covariates (default 4)}
}
\value{
matrix of base covariates
}
\description{
Generate base covariates
}
