% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{generate_site_treatment_assignments}
\alias{generate_site_treatment_assignments}
\title{Generate site membership and treatment assignments (vectorized)}
\usage{
generate_site_treatment_assignments(n, site_probs, K)
}
\arguments{
\item{n}{number of observations}

\item{site_probs}{site probability list}

\item{K}{number of source sites}
}
\value{
list with R (site membership) and A (treatment) vectors
}
\description{
Generate site membership and treatment assignments (vectorized)
}
