% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{glm_integer_codes}
\alias{glm_integer_codes}
\alias{FAMILY_GAUSSIAN}
\alias{FAMILY_BINOMIAL}
\alias{LINK_IDENTITY}
\alias{LINK_LOGIT}
\title{Integer codes for supported GLM families and links}
\format{
An object of class \code{integer} of length 1.

An object of class \code{integer} of length 1.

An object of class \code{integer} of length 1.

An object of class \code{integer} of length 1.
}
\usage{
FAMILY_GAUSSIAN

FAMILY_BINOMIAL

LINK_IDENTITY

LINK_LOGIT
}
\value{
Integer scalar.
}
\description{
These constants mirror the enumerations used by the compiled nuisance-model
kernels.
}
\keyword{datasets}
