% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{HESSIAN_FLOOR}
\alias{HESSIAN_FLOOR}
\title{Hessian floor (matches C++ HESSIAN_FLOOR)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
HESSIAN_FLOOR
}
\description{
Hessian floor (matches C++ HESSIAN_FLOOR)
}
\keyword{datasets}
