% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{INDEPENDENT_SHIFT_SD}
\alias{INDEPENDENT_SHIFT_SD}
\title{Shift SD multiplier for independent site allocation covariate shifts
Controls magnitude of site-specific mean shifts: delta ~ N(0, INDEPENDENT_SHIFT_SD * shift_strength)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
INDEPENDENT_SHIFT_SD
}
\description{
Shift SD multiplier for independent site allocation covariate shifts
Controls magnitude of site-specific mean shifts: delta ~ N(0, INDEPENDENT_SHIFT_SD * shift_strength)
}
\keyword{datasets}
