% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{LAMBDA_GRID_SIZE_STANDARD}
\alias{LAMBDA_GRID_SIZE_STANDARD}
\title{Lambda grid size for standard CV selection}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
LAMBDA_GRID_SIZE_STANDARD
}
\description{
Lambda grid size for standard CV selection
}
\keyword{datasets}
