% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{LAMBDA_MIN}
\alias{LAMBDA_MIN}
\title{Minimum lambda for weight optimization}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
LAMBDA_MIN
}
\description{
Minimum lambda for weight optimization
}
\keyword{datasets}
