% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{MAX_ITER_DEFAULT}
\alias{MAX_ITER_DEFAULT}
\title{Maximum iterations for coordinate descent optimization}
\format{
An object of class \code{integer} of length 1.
}
\usage{
MAX_ITER_DEFAULT
}
\description{
Maximum iterations for coordinate descent optimization
}
\keyword{datasets}
