% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{N_CV_FOLDS_LAMBDA}
\alias{N_CV_FOLDS_LAMBDA}
\title{Number of CV folds for internal lambda selection in C++ nuisance models.
Separate from N_FOLDS_DEFAULT (outer cross-fitting) to allow independent tuning.}
\format{
An object of class \code{integer} of length 1.
}
\usage{
N_CV_FOLDS_LAMBDA
}
\description{
Number of CV folds for internal lambda selection in C++ nuisance models.
Separate from N_FOLDS_DEFAULT (outer cross-fitting) to allow independent tuning.
}
\keyword{datasets}
