% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{N_FOLDS_DEFAULT}
\alias{N_FOLDS_DEFAULT}
\title{Default number of folds for the outer cross-fitting loop (K_f in paper)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
N_FOLDS_DEFAULT
}
\description{
Default number of folds for the outer cross-fitting loop (K_f in paper)
}
\keyword{datasets}
