% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numerical_utils.R
\name{op-null-default}
\alias{op-null-default}
\alias{\%||\%}
\title{Null-coalescing operator (backport for R < 4.1.0)}
\usage{
x \%||\% y
}
\arguments{
\item{x}{Value to test.}

\item{y}{Default value when x is NULL.}
}
\value{
x if not NULL, else y.
}
\description{
Returns x if it is not NULL, otherwise returns y.
}
