% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{OUTCOME_PRED_LOWER}
\alias{OUTCOME_PRED_LOWER}
\title{Outcome prediction clipping bounds (prevent log(0) in binomial deviance)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
OUTCOME_PRED_LOWER
}
\description{
Outcome prediction clipping bounds (prevent log(0) in binomial deviance)
}
\keyword{datasets}
