% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{PARAM_MAX}
\alias{PARAM_MAX}
\title{Parameter clipping bound (matches C++ PARAM_MAX)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
PARAM_MAX
}
\description{
Parameter clipping bound (matches C++ PARAM_MAX)
}
\keyword{datasets}
