% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{POSITIVITY_LOWER}
\alias{POSITIVITY_LOWER}
\title{Positivity lower bound for propensity score clipping in data generation}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
POSITIVITY_LOWER
}
\description{
Positivity lower bound for propensity score clipping in data generation
}
\keyword{datasets}
