% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{POSITIVITY_UPPER}
\alias{POSITIVITY_UPPER}
\title{Positivity upper bound for propensity score clipping in data generation}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
POSITIVITY_UPPER
}
\description{
Positivity upper bound for propensity score clipping in data generation
}
\keyword{datasets}
