% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{PROP_SCORE_LOWER}
\alias{PROP_SCORE_LOWER}
\title{Propensity score clipping bounds (prevent extreme IPW weights)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
PROP_SCORE_LOWER
}
\description{
Propensity score clipping bounds (prevent extreme IPW weights)
}
\keyword{datasets}
