% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{RIDGE_DEFAULT}
\alias{RIDGE_DEFAULT}
\title{Ridge regularization for matrix inversion}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
RIDGE_DEFAULT
}
\description{
Ridge regularization for matrix inversion
}
\keyword{datasets}
