% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{ROCE_NOISE_SD_DEFAULT}
\alias{ROCE_NOISE_SD_DEFAULT}
\title{Default noise standard deviation for continuous outcomes in RoCE DGP}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
ROCE_NOISE_SD_DEFAULT
}
\description{
Default noise standard deviation for continuous outcomes in RoCE DGP
}
\keyword{datasets}
