% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkpoint.R
\name{signal_checkpoint_saved}
\alias{signal_checkpoint_saved}
\title{Signal that a checkpoint has been saved (for shell-script coordination)}
\usage{
signal_checkpoint_saved(saved_signal_file)
}
\arguments{
\item{saved_signal_file}{Path to the saved-signal file.}
}
\description{
Signal that a checkpoint has been saved (for shell-script coordination)
}
