% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{split_data_by_site}
\alias{split_data_by_site}
\title{Split data by site}
\usage{
split_data_by_site(data)
}
\arguments{
\item{data}{full dataset}
}
\value{
list with data for each site
}
\description{
Split data by site
}
