% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{TOL_DEFAULT}
\alias{TOL_DEFAULT}
\title{Convergence tolerance for optimization}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
TOL_DEFAULT
}
\description{
Convergence tolerance for optimization
}
\keyword{datasets}
