% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{VALID_CONFIGS}
\alias{VALID_CONFIGS}
\title{Valid configuration identifiers}
\format{
An object of class \code{character} of length 4.
}
\usage{
VALID_CONFIGS
}
\description{
Valid configuration identifiers
}
\keyword{datasets}
