% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{VALID_ESTIMAND_TYPES}
\alias{VALID_ESTIMAND_TYPES}
\title{Valid estimand types}
\format{
An object of class \code{character} of length 2.
}
\usage{
VALID_ESTIMAND_TYPES
}
\description{
Valid estimand types
}
\keyword{datasets}
