% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{VALID_GLM_FAMILIES}
\alias{VALID_GLM_FAMILIES}
\title{Valid GLM family names (glmnet-style API)}
\format{
An object of class \code{character} of length 2.
}
\usage{
VALID_GLM_FAMILIES
}
\description{
Valid GLM family names (glmnet-style API)
}
\keyword{datasets}
