% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{VALID_HETEROGENEITY_TYPES}
\alias{VALID_HETEROGENEITY_TYPES}
\title{Valid heterogeneity types}
\format{
An object of class \code{character} of length 4.
}
\usage{
VALID_HETEROGENEITY_TYPES
}
\description{
Valid heterogeneity types
}
\keyword{datasets}
