% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{VALID_OUTCOME_TYPES}
\alias{VALID_OUTCOME_TYPES}
\title{Valid outcome types}
\format{
An object of class \code{character} of length 2.
}
\usage{
VALID_OUTCOME_TYPES
}
\description{
Valid outcome types
}
\keyword{datasets}
