% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{VALID_SITE_ALLOCATIONS}
\alias{VALID_SITE_ALLOCATIONS}
\title{Valid site allocation methods}
\format{
An object of class \code{character} of length 6.
}
\usage{
VALID_SITE_ALLOCATIONS
}
\description{
Valid site allocation methods
}
\keyword{datasets}
