% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{VALID_TRANSFORM_TYPES}
\alias{VALID_TRANSFORM_TYPES}
\title{Valid covariate transformation types}
\format{
An object of class \code{character} of length 3.
}
\usage{
VALID_TRANSFORM_TYPES
}
\description{
Valid covariate transformation types
}
\keyword{datasets}
