% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{VARIANCE_MIN}
\alias{VARIANCE_MIN}
\title{Minimum variance floor (prevent division by zero)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
VARIANCE_MIN
}
\description{
Minimum variance floor (prevent division by zero)
}
\keyword{datasets}
