% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{WEIGHT_MIN}
\alias{WEIGHT_MIN}
\title{Density ratio weight bounds (matches C++ WEIGHT_MIN/MAX)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
WEIGHT_MIN
}
\description{
Density ratio weight bounds (matches C++ WEIGHT_MIN/MAX)
}
\keyword{datasets}
