% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{WEIGHT_OPT_MAX_ITER}
\alias{WEIGHT_OPT_MAX_ITER}
\title{Maximum iterations for weight optimization (optimize_weights)}
\format{
An object of class \code{integer} of length 1.
}
\usage{
WEIGHT_OPT_MAX_ITER
}
\description{
Maximum iterations for weight optimization (optimize_weights)
}
\keyword{datasets}
