% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{WEIGHT_OPT_TOL}
\alias{WEIGHT_OPT_TOL}
\title{Convergence tolerance for weight optimization (optimize_weights)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
WEIGHT_OPT_TOL
}
\description{
Convergence tolerance for weight optimization (optimize_weights)
}
\keyword{datasets}
