% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{Z_ALPHA_05}
\alias{Z_ALPHA_05}
\title{Z-score for 95\% confidence intervals}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
Z_ALPHA_05
}
\description{
Z-score for 95\% confidence intervals
}
\keyword{datasets}
